- Move all service class objects into services directory (from models) - Refactor aliasing and query expansion methods from SolrQuery into ExpandQuery service class - Add initial spec for ExpandQuery - Add SesQuery spec
