+++ app/services/associated_objects.rb
