+++ app/services/associated_objects_for_object_view.rb
