+++ app/services/associated_objects_for_search_results.rb
