+++ app/services/hierarchy_builder.rb
