+++ app/services/solr_multi_query.rb
