+++ app/services/solr_query.rb
