+++ app/services/solr_query_wrapper.rb
