+++ spec/services/facet_spec.rb
