+++ spec/services/solr_multi_query_spec.rb
