+++ spec/services/solr_query_spec.rb
