+++ spec/services/solr_query_wrapper_spec.rb
