+++ spec/services/solr_search_spec.rb
